# model: add time tracking to Matcher