--- a/model/labels/regexp.go
+++ b/model/labels/regexp.go
@@ -24,6 +24,7 @@ import (
 	"github.com/dgraph-io/ristretto"
 	"github.com/grafana/regexp"
 	"github.com/grafana/regexp/syntax"
+	"go.uber.org/atomic"
 	"golang.org/x/text/unicode/norm"
 )
 
@@ -67,6 +68,10 @@ type FastRegexMatcher struct {
 
 	// matchString is the "compiled" function to run by MatchString().
 	matchString func(string) bool
+
+	// time tracking
+	matchesWallClockDuration *atomic.Duration
+	sampler                  int64
 }
 
 func NewFastRegexMatcher(v string) (*FastRegexMatcher, error) {
@@ -343,6 +348,28 @@ func tooManyMatches(matches []string, added ...string) bool {
 }
 
 func (m *FastRegexMatcher) MatchString(s string) bool {
+	if m.matchesWallClockDuration != nil {
+		// here we allow for data races.
+		m.sampler++
+		tick := m.sampler
+		// Check if it's a power of two; if yes, then we record.
+		// This way we amortize the cost of recording latency.
+		if tick&(tick-1) == 0 {
+			defer func(start time.Time, durationAtStart time.Duration) {
+				// Use an amortized time.Now() to amortize the cost.
+				// amortizedNow has a lower precision, but is cheaper to access than time.Now()
+				// Using time.Now() makes Matches() TODO
+				thisInvocationDuration := amortizedNow.Load().Sub(start)
+				previousInvocationsDuration := thisInvocationDuration * time.Duration(max(1, tick/2))
+
+				// If the CAS fails, then some other goroutine updated it.
+				// In that case either of these goroutines wasn't the "right" goroutine to update it because of a race on `sampler`
+				// We accept that data race so that we don't pay the price of synchronisa tion on sampler.
+				// Using an atomic makes Matches() 30% slower in benchamrks. TODO
+				_ = m.matchesWallClockDuration.CompareAndSwap(durationAtStart, durationAtStart+previousInvocationsDuration)
+			}(amortizedNow.Load(), m.matchesWallClockDuration.Load())
+		}
+	}
 	return m.matchString(s)
 }