--- a/model/labels/regexp.go
+++ b/model/labels/regexp.go
@@ -72,7 +72,6 @@ type FastRegexMatcher struct {
 	// time tracking
 	matchesWallClockDuration *atomic.Duration
 	sampler                  int64
-	sampleRate               int64
 }
 
 func NewFastRegexMatcher(v string) (*FastRegexMatcher, error) {
@@ -350,15 +349,15 @@ func tooManyMatches(matches []string, added ...string) bool {
 
 func (m *FastRegexMatcher) MatchString(s string) bool {
 	if m.matchesWallClockDuration != nil {
-		// here we allow for data races.
 		m.sampler++
 		tick := m.sampler
 		// Check if it's a power of two; if yes, then we record.
 		// This way we amortize the cost of recording latency.
-		if tick&m.sampleRate == 0 {
+		// Sampling one of every 64 executions makes Matches() ~25% slower in benchmarks.
+		if tick&(tick-1) == 0 {
 			defer func(start time.Time, durationAtStart time.Duration) {
 				thisInvocationDuration := time.Since(start)
-				previousInvocationsDuration := thisInvocationDuration * time.Duration(tick*m.sampleRate)
+				previousInvocationsDuration := thisInvocationDuration * time.Duration(max(1, tick/2))
 
 				// If the CAS fails, then some other goroutine updated it.
 				// In that case either of these goroutines wasn't the "right" goroutine to update it because of a race on `sampler`