> This is an internal version of https://github.com/prometheus/prometheus/pull/16853 so that we can move forward with the work while waiting for the upstream PR to be accepted.


### Background

In grafana/mimir we sometimes see very expensive regular expressions. Unfortunately, apart from profiling and profiling recording rules, there is no insight into how impactful these regular expressions are.

### What this PR does

Adds `NewMatcherWithTimeTracker` which increments an `*atomic.Duration` when the matcher is executed. This helps to put the time of regular expressions into perspective. For example, we can use it to measure the portion of time of serving a query which went to running regular expressions.

### Sampling executions

This PR adds sampling to the execution tracking. The reason is that tracking every execution can lead to significant time spent in `time.Now()` and `time.Since()` and atomic operations. Sampling linearly still has the same effects.


Instead of sampling every 64 or 128 invocation, we sample the 1st, 2nd, 4th, 8th, 16th, 32nd... This helps a lot. We get roughly 20% slowdown with linear sampling compared to exponential sampling.


### Data races

Sampling in a thread-safe way is expensive, so I chose to not support concurrent uses of matchers. In reality we don't use matchers concurrently in a single request in Mimir, so this shouldn't be a problem.

They're not goroutine-safe, but not across cached versions of the matchers: retrieving the same matcher from the cache and using it concurrently is safe. Using the same matcher retrieved only once from the cache concurrently is not safe.



<!--
    Please give your PR a title in the form "area: short description".  For example "tsdb: reduce disk usage by 95%"

    If your PR is to fix an issue, put "Fixes #issue-number" in the description.

    Don't forget!

    - Please sign CNCF's Developer Certificate of Origin and sign-off your commits by adding the -s / --signoff flag to `git commit`. See https://github.com/apps/dco for more information.

    - If the PR adds or changes a behaviour or fixes a bug of an exported API it would need a unit/e2e test.

    - Where possible use only exported APIs for tests to simplify the review and make it as close as possible to an actual library usage.

    - Performance improvements would need a benchmark test to prove it.

    - All exposed objects should have a comment.

    - All comments should start with a capital letter and end with a full stop.
 -->


<!--
    Please give your PR a title in the form "area: short description".  For example "tsdb: reduce disk usage by 95%"

    If your PR is to fix an issue, put "Fixes #issue-number" in the description.

    Don't forget!

    - Please sign CNCF's Developer Certificate of Origin and sign-off your commits by adding the -s / --signoff flag to `git commit`. See https://github.com/apps/dco for more information.

    - If the PR adds or changes a behaviour or fixes a bug of an exported API it would need a unit/e2e test.

    - Where possible use only exported APIs for tests to simplify the review and make it as close as possible to an actual library usage.

    - Performance improvements would need a benchmark test to prove it.

    - All exposed objects should have a comment.

    - All comments should start with a capital letter and end with a full stop.
 -->