@dimitarvdimitrov 2025-07-24
> > Do you have an idea of how expensive it is to measure, e.g., all executions vs every 64th?

So I actually decided to run some benchmarks to prove the point of this, that having it hardcoded is better and to show the impact of time tracking. 

I pushed the benchmark and the results 665ecab6d66c8a11ee6fc7c18c200bbfe5c67c90 

### Hardcoded power of 2 constant vs dynamic variable

It looks like having it configurable versus having it hardcoded doesn't have an effect on regular expression evaluations.

```
benchstat -filter '/withDynamicSampling:(false OR true)' -row /sampleRate -col /withDynamicSampling benchmarks.txt  2>/dev/null
goos: darwin
goarch: arm64
pkg: github.com/prometheus/prometheus/model/labels
cpu: Apple M1 Pro
        │     false     │                 true                  │
        │    sec/op     │    sec/op      vs base                │
64        166.7n ± 5% ¹   165.2n ± 5% ¹       ~ (p=0.592 n=378)
128       159.2n ± 6% ¹   157.2n ± 6% ¹       ~ (p=0.183 n=378)
geomean   162.9n          161.2n         -1.02%
¹ benchmarks vary in .fullname
```

I think this just makes it obvious that we should have it configurable.


### Time tracking vs without

It does look like time-tracking is adding 34% overhead to regular expression matching :(


```
benchstat -filter '/withTimeTracking:(false OR true)' -row . -col /sampleRate,/withTimeTracking benchmarks.txt  2>/dev/null
goos: darwin
goarch: arm64
pkg: github.com/prometheus/prometheus/model/labels
cpu: Apple M1 Pro
 │      64       │                  128                  │                                    │
 │     true      │                 true                  │               false                │
 │    sec/op     │    sec/op      vs base                │    sec/op      vs base             │
   166.8n ± 5% ¹   158.5n ± 5% ¹  -4.98% (p=0.000 n=770)   109.4n ± 9% ¹  -34.44% (n=770+378)
¹ benchmarks vary in .fullname
```



The overhead is more than I expected and I think it might be significant in deployments where regular expressions take up a double-digit % of CPU