@dimitarvdimitrov 2025-09-10
this is too expensive. A more practical approach is using labels on pprof profiles and profiles recording rules with pyroscope (see grafana/mimir#12404)